+++ forui/example/lib/portal/visualizer.dart
@@ -44,7 +44,7 @@ class _VisualizerState extends State<Visualizer> {
           label: const Text('Shift'),
           onChange: (closure) => setState(() => _shift = closure ?? _shift),
           format: (closure) => _shifts[closure]!,
-          children: [for (final MapEntry(:key, :value) in _shifts.entries) FSelectItem(value, key)],
+          children: [for (final MapEntry(:key, :value) in _shifts.entries) FSelectItem(title: Text(value), value:  key)],
         ),
         const SizedBox(height: 16),
         _Settings(
@@ -233,7 +233,7 @@ class _Settings extends StatelessWidget {
               Alignment.bottomCenter,
               Alignment.bottomRight,
             ])
-              FSelectItem(anchor.toString().replaceAll('Alignment.', ''), anchor),
+              FSelectItem(title: Text(anchor.toString().replaceAll('Alignment.', '')), value: anchor),
           ],
         ),
       ],