+++ forui/lib/src/widgets/select/select_item.dart
@@ -40,6 +40,8 @@ class FSelectSection<T> extends StatelessWidget with FSelectItemMixin {
   final List<FSelectItem<T>> children;
 
   /// Creates a [FSelectSection].
+  ///
+  /// For a convenient way to create a section, use [FSelectSection.fromMap].
   const FSelectSection({
     required this.label,
     required this.children,
@@ -49,7 +51,9 @@ class FSelectSection<T> extends StatelessWidget with FSelectItemMixin {
     super.key,
   });
 
-  /// Creates a [FSelectSection] from the given items.
+  /// Creates a [FSelectSection] from the given [items].
+  ///
+  /// For more control over the items' appearances, use [FSelectSection].
   FSelectSection.fromMap({
     required Widget label,
     required Map<String, T> items,
@@ -59,7 +63,7 @@ class FSelectSection<T> extends StatelessWidget with FSelectItemMixin {
     Key? key,
   }) : this(
          label: label,
-         children: [for (final e in items.entries) FSelectItem<T>(e.key, e.value)],
+         children: [for (final e in items.entries) FSelectItem<T>(title: Text(e.key), value: e.value)],
          style: style,
          enabled: enabled,
          divider: divider,
@@ -71,7 +75,7 @@ class FSelectSection<T> extends StatelessWidget with FSelectItemMixin {
     final content = ContentData.of<T>(context);
     final enabled = this.enabled ?? content.enabled;
     final style = this.style?.call(content.style) ?? content.style;
-    final itemStyle = style.itemStyle.toFItemStyle(context);
+    final itemStyle = style.itemStyle.call(context);
 
     return ContentData<T>(
       style: style,
@@ -151,7 +155,7 @@ class FSelectSectionStyle with Diagnosticable, _$FSelectSectionStyleFunctions {
 
   /// The section's items' style.
   @override
-  final FSelectItemStyle itemStyle;
+  final FItemStyle itemStyle;
 
   /// Creates a [FSelectSectionStyle].
   FSelectSectionStyle({
@@ -163,34 +167,77 @@ class FSelectSectionStyle with Diagnosticable, _$FSelectSectionStyleFunctions {
   });
 
   /// Creates a [FSelectSectionStyle] that inherits its properties.
-  FSelectSectionStyle.inherit({required FColors colors, required FStyle style, required FTypography typography})
-    : this(
-        labelTextStyle: FWidgetStateMap({
-          WidgetState.disabled: typography.sm.copyWith(
-            color: colors.disable(colors.primary),
-            fontWeight: FontWeight.w600,
+  factory FSelectSectionStyle.inherit({
+    required FColors colors,
+    required FStyle style,
+    required FTypography typography,
+  }) {
+    const padding = EdgeInsetsDirectional.only(start: 11, top: 7.5, bottom: 7.5, end: 6);
+    final iconStyle = FWidgetStateMap({
+      WidgetState.disabled: IconThemeData(color: colors.disable(colors.primary), size: 15),
+      WidgetState.any: IconThemeData(color: colors.primary, size: 15),
+    });
+    final textStyle = FWidgetStateMap({
+      WidgetState.disabled: typography.sm.copyWith(color: colors.disable(colors.primary)),
+      WidgetState.any: typography.sm.copyWith(color: colors.primary),
+    });
+
+    return FSelectSectionStyle(
+      labelTextStyle: FWidgetStateMap({
+        WidgetState.disabled: typography.sm.copyWith(
+          color: colors.disable(colors.primary),
+          fontWeight: FontWeight.w600,
+        ),
+        WidgetState.any: typography.sm.copyWith(color: colors.primary, fontWeight: FontWeight.w600),
+      }),
+      dividerColor: FWidgetStateMap.all(colors.border),
+      dividerWidth: style.borderWidth,
+      itemStyle: FItemStyle(
+        backgroundColor: FWidgetStateMap.all(null),
+        decoration: FWidgetStateMap({
+          ~WidgetState.disabled & (WidgetState.focused | WidgetState.hovered | WidgetState.pressed): BoxDecoration(
+            color: colors.secondary,
+            borderRadius: style.borderRadius,
           ),
-          WidgetState.any: typography.sm.copyWith(color: colors.primary, fontWeight: FontWeight.w600),
         }),
-        dividerColor: FWidgetStateMap.all(colors.border),
-        dividerWidth: style.borderWidth,
-        itemStyle: FSelectItemStyle.inherit(colors: colors, style: style, typography: typography),
-      );
+        contentStyle: FItemContentStyle.inherit(colors: colors, typography: typography).copyWith(
+          padding: padding,
+          prefixIconStyle: iconStyle,
+          prefixIconSpacing: 10,
+          titleTextStyle: textStyle,
+          titleSpacing: 4,
+          subtitleTextStyle: FWidgetStateMap({
+            WidgetState.disabled: typography.xs.copyWith(color: colors.disable(colors.mutedForeground)),
+            WidgetState.any: typography.xs.copyWith(color: colors.mutedForeground),
+          }),
+          suffixIconStyle: FWidgetStateMap({
+            WidgetState.disabled: IconThemeData(color: colors.disable(colors.primary), size: 15),
+            WidgetState.any: IconThemeData(color: colors.primary, size: 15),
+          }),
+        ),
+        rawItemContentStyle: FRawItemContentStyle(
+          padding: padding,
+          prefixIconStyle: iconStyle,
+          childTextStyle: textStyle,
+        ),
+        tappableStyle: style.tappableStyle.copyWith(bounceTween: FTappableStyle.noBounceTween),
+        focusedOutlineStyle: null,
+      ),
+    );
+  }
 }
 
 /// A selectable item in a [FSelect] that can optionally be nested in a [FSelectSection].
-class FSelectItem<T> extends StatefulWidget with FSelectItemMixin {
-  static Widget? _defaultSuffixBuilder(BuildContext _, bool selected) => selected ? const Icon(FIcons.check) : null;
-
-  /// The style. Defaults to the [FSelectItemStyle] inherited from the parent [FSelectSection] or [FSelect].
+abstract class FSelectItem<T> extends StatefulWidget with FSelectItemMixin {
+  /// The style. Defaults to the [FItemStyle] inherited from the parent [FSelectSection] or [FSelect].
   ///
   /// ## CLI
   /// To generate and customize this style:
   ///
   /// ```shell
-  /// dart run forui style create select-item
+  /// dart run forui style create select-section
   /// ```
-  final FSelectItemStyle Function(FSelectItemStyle)? style;
+  final FItemStyle Function(FItemStyle)? style;
 
   /// The value.
   final T value;
@@ -203,34 +250,32 @@ class FSelectItem<T> extends StatefulWidget with FSelectItemMixin {
   /// A prefix.
   final Widget? prefix;
 
-  /// The subtitle.
-  final Widget? subtitle;
+  /// Creates a [FSelectItem] with a custom [title] and value.
+  ///
+  /// For even more control over the item's appearance, use [FSelectItem.raw].
+  const factory FSelectItem({
+    required Widget title,
+    required T value,
+    FItemStyle Function(FItemStyle)? style,
+    bool? enabled,
+    Widget? prefix,
+    Widget? subtitle,
+    // ignore: avoid_positional_boolean_parameters
+    Widget? Function(BuildContext, bool) suffixBuilder,
+    Key? key,
+  }) = _SelectItem<T>;
 
-  /// The child.
-  final Widget title;
+  /// Creates a [FSelectItem] with a raw layout.
+  const factory FSelectItem.raw({
+    required Widget child,
+    required T value,
+    FItemStyle Function(FItemStyle)? style,
+    bool? enabled,
+    Widget? prefix,
+    Key? key,
+  }) = _RawSelectItem<T>;
 
-  /// The icon displayed when the item is selected. Defaults to a check icon.
-  // ignore: avoid_positional_boolean_parameters
-  final Widget? Function(BuildContext, bool) suffixBuilder;
-
-  /// Creates a [FSelectItem].
-  FSelectItem(String text, T value, {bool? enabled, Key? key})
-    : this.from(key: key, value: value, enabled: enabled, title: Text(text));
-
-  /// Creates a [FSelectItem] with a custom child widget.
-  FSelectItem.from({
-    required this.value,
-    required this.title,
-    this.style,
-    this.enabled,
-    this.prefix,
-    this.subtitle,
-    this.suffixBuilder = _defaultSuffixBuilder,
-    super.key,
-  });
-
-  @override
-  State<FSelectItem<T>> createState() => _FSelectItemState<T>();
+  const FSelectItem._({required this.value, this.style, this.enabled, this.prefix, super.key});
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -238,12 +283,11 @@ class FSelectItem<T> extends StatefulWidget with FSelectItemMixin {
     properties
       ..add(DiagnosticsProperty('style', style))
       ..add(DiagnosticsProperty('value', value))
-      ..add(FlagProperty('enabled', value: enabled, ifTrue: 'enabled', ifFalse: 'disabled'))
-      ..add(ObjectFlagProperty.has('suffixBuilder', suffixBuilder));
+      ..add(FlagProperty('enabled', value: enabled, ifTrue: 'enabled', ifFalse: 'disabled'));
   }
 }
 
-class _FSelectItemState<T> extends State<FSelectItem<T>> {
+abstract class _State<W extends FSelectItem<T>, T> extends State<W> {
   late final _focus = FocusNode(debugLabel: widget.value.toString());
 
   @override
@@ -265,155 +309,122 @@ class _FSelectItemState<T> extends State<FSelectItem<T>> {
   }
 
   @override
-  Widget build(BuildContext context) {
-    final InheritedSelectController(:popover, :contains, :focus, :onPress) = InheritedSelectController.of<T>(context);
-    final content = ContentData.of<T>(context);
-
-    final enabled = widget.enabled ?? content.enabled;
-    final selected = contains(widget.value);
-    final focused = focus(widget.value);
-    final style = widget.style?.call(content.style.itemStyle).toFItemStyle(context);
-
-    return FItem(
-      style: style?.call,
-      enabled: enabled,
-      selected: selected,
-      autofocus: focused || content.first,
-      focusNode: _focus,
-      onPress: () => onPress(widget.value),
-      onHoverChange: (hover) {
-        if (popover.status.isCompleted) {
-          hover ? _focus.requestFocus() : _focus.unfocus();
-        }
-      },
-      prefix: widget.prefix,
-      title: widget.title,
-      subtitle: widget.subtitle,
-      suffix: widget.suffixBuilder(context, selected),
-    );
-  }
-
-  @override
   void dispose() {
     _focus.dispose();
     super.dispose();
   }
+
+  @override
+  Widget build(BuildContext context) {
+    final InheritedSelectController(:popover, :contains, :focus, :onPress) = InheritedSelectController.of<T>(context);
+    final content = ContentData.of<T>(context);
+
+    return _item(
+      context,
+      widget.enabled ?? content.enabled,
+      contains(widget.value),
+      focus(widget.value) || content.first,
+      () => onPress(widget.value),
+      (hover) {
+        if (popover.status.isCompleted) {
+          hover ? _focus.requestFocus() : _focus.unfocus();
+        }
+      },
+    );
+  }
+
+  Widget _item(
+    BuildContext context,
+    bool enabled,
+    bool selected,
+    bool focused,
+    VoidCallback onPress,
+    ValueChanged<bool> onHover,
+  );
 }
 
-/// A [FSelectItem]'s style.
-class FSelectItemStyle with Diagnosticable, _$FSelectItemStyleFunctions {
-  /// The margin around the item. Defaults to `EdgeInsets.symmetric(horizontal: 4, vertical: 2)`.
+class _SelectItem<T> extends FSelectItem<T> {
+  static Widget? _defaultSuffixBuilder(BuildContext _, bool selected) => selected ? const Icon(FIcons.check) : null;
+
+  final Widget? subtitle;
+  final Widget title;
+
+  // ignore: avoid_positional_boolean_parameters
+  final Widget? Function(BuildContext, bool) suffixBuilder;
+
+  const _SelectItem({
+    required this.title,
+    required super.value,
+    this.subtitle,
+    this.suffixBuilder = _defaultSuffixBuilder,
+    super.style,
+    super.enabled,
+    super.prefix,
+    super.key,
+  }) : super._();
+
   @override
-  final EdgeInsetsGeometry margin;
+  State<_SelectItem<T>> createState() => _SelectItemState<T>();
 
-  /// The padding around the item. Defaults to `EdgeInsetsDirectional.only(start: 15, top: 7.5, bottom: 7.5, end: 10)`.
   @override
-  final EdgeInsetsGeometry padding;
-
-  /// The decoration.
-  ///
-  /// {@macro forui.foundation.doc_templates.WidgetStates.selectable}
-  @override
-  final FWidgetStateMap<BoxDecoration?> decoration;
-
-  /// The icon style for an item's prefix.
-  ///
-  /// {@macro forui.foundation.doc_templates.WidgetStates.selectable}
-  @override
-  final FWidgetStateMap<IconThemeData> prefixIconStyle;
-
-  // The horizontal spacing between the prefix icon and title and the subtitle. Defaults to 10.
-  @override
-  final double prefixIconSpacing;
-
-  /// The default text style for the title.
-  ///
-  /// {@macro forui.foundation.doc_templates.WidgetStates.selectable}
-  @override
-  final FWidgetStateMap<TextStyle> titleTextStyle;
-
-  /// The vertical spacing between the title and the subtitle. Defaults to 4.
-  @override
-  final double titleSpacing;
-
-  /// The default text style for the subtitle.
-  ///
-  /// {@macro forui.foundation.doc_templates.WidgetStates.selectable}
-  @override
-  final FWidgetStateMap<TextStyle> subtitleTextStyle;
-
-  /// The icon style for an item's suffix.
-  ///
-  /// {@macro forui.foundation.doc_templates.WidgetStates.selectable}
-  @override
-  final FWidgetStateMap<IconThemeData> suffixIconStyle;
-
-  /// The tappable style for the item.
-  @override
-  final FTappableStyle tappableStyle;
-
-  /// Creates a [FSelectItemStyle].
-  FSelectItemStyle({
-    required this.decoration,
-    required this.prefixIconStyle,
-    required this.titleTextStyle,
-    required this.subtitleTextStyle,
-    required this.suffixIconStyle,
-    required this.tappableStyle,
-    this.margin = const EdgeInsets.symmetric(horizontal: 4, vertical: 2),
-    this.padding = const EdgeInsetsDirectional.only(start: 11, top: 7.5, bottom: 7.5, end: 6),
-    this.prefixIconSpacing = 10,
-    this.titleSpacing = 4,
-  });
-
-  /// Creates a [FSelectItemStyle] that inherits its properties.
-  FSelectItemStyle.inherit({required FColors colors, required FStyle style, required FTypography typography})
-    : this(
-        decoration: FWidgetStateMap({
-          ~WidgetState.disabled & (WidgetState.focused | WidgetState.hovered | WidgetState.pressed): BoxDecoration(
-            color: colors.secondary,
-            borderRadius: style.borderRadius,
-          ),
-        }),
-        prefixIconStyle: FWidgetStateMap({
-          WidgetState.disabled: IconThemeData(color: colors.disable(colors.primary), size: 15),
-          WidgetState.any: IconThemeData(color: colors.primary, size: 15),
-        }),
-        titleTextStyle: FWidgetStateMap({
-          WidgetState.disabled: typography.sm.copyWith(color: colors.disable(colors.primary)),
-          WidgetState.any: typography.sm.copyWith(color: colors.primary),
-        }),
-        subtitleTextStyle: FWidgetStateMap({
-          WidgetState.disabled: typography.xs.copyWith(color: colors.disable(colors.mutedForeground)),
-          WidgetState.any: typography.xs.copyWith(color: colors.mutedForeground),
-        }),
-        suffixIconStyle: FWidgetStateMap({
-          WidgetState.disabled: IconThemeData(color: colors.disable(colors.primary), size: 15),
-          WidgetState.any: IconThemeData(color: colors.primary, size: 15),
-        }),
-        tappableStyle: style.tappableStyle.copyWith(bounceTween: FTappableStyle.noBounceTween),
-      );
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(ObjectFlagProperty.has('suffixBuilder', suffixBuilder));
+  }
 }
 
-@internal
-extension InternalFSelectItemStyle on FSelectItemStyle {
-  FItemStyle toFItemStyle(BuildContext context) => FItemStyle(
-    backgroundColor: FWidgetStateMap.all(null),
-    decoration: decoration,
-    margin: margin,
-    contentStyle: FItemContentStyle.inherit(colors: context.theme.colors, typography: context.theme.typography)
-        .copyWith(
-          padding: padding,
-          prefixIconStyle: prefixIconStyle,
-          prefixIconSpacing: prefixIconSpacing,
-          titleTextStyle: titleTextStyle,
-          titleSpacing: titleSpacing,
-          subtitleTextStyle: subtitleTextStyle,
-          suffixIconStyle: suffixIconStyle,
-        ),
-    rawItemContentStyle: context.theme.itemStyle.rawItemContentStyle,
-    // This isn't ever used.
-    tappableStyle: tappableStyle,
-    focusedOutlineStyle: null,
+class _SelectItemState<T> extends _State<_SelectItem<T>, T> {
+  @override
+  Widget _item(
+    BuildContext context,
+    bool enabled,
+    bool selected,
+    bool focused,
+    VoidCallback onPress,
+    ValueChanged<bool> onHover,
+  ) => FItem(
+    style: widget.style?.call,
+    enabled: enabled,
+    selected: selected,
+    autofocus: focused,
+    focusNode: _focus,
+    onPress: onPress,
+    onHoverChange: onHover,
+    prefix: widget.prefix,
+    title: widget.title,
+    subtitle: widget.subtitle,
+    suffix: widget.suffixBuilder(context, selected),
+  );
+}
+
+class _RawSelectItem<T> extends FSelectItem<T> {
+  final Widget child;
+
+  const _RawSelectItem({required this.child, required super.value, super.style, super.enabled, super.prefix, super.key})
+    : super._();
+
+  @override
+  _RawSelectItemState<T> createState() => _RawSelectItemState<T>();
+}
+
+class _RawSelectItemState<T> extends _State<_RawSelectItem<T>, T> {
+  @override
+  Widget _item(
+    BuildContext context,
+    bool enabled,
+    bool selected,
+    bool focused,
+    VoidCallback onPress,
+    ValueChanged<bool> onHover,
+  ) => FItem.raw(
+    style: widget.style?.call,
+    enabled: enabled,
+    selected: selected,
+    autofocus: focused,
+    focusNode: _focus,
+    onPress: onPress,
+    onHoverChange: onHover,
+    prefix: widget.prefix,
+    child: widget.child,
   );
 }