+++ forui/lib/src/widgets/select/single/select.dart
@@ -321,7 +321,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
       contentDivider: contentDivider,
       initialValue: initialValue,
       key: key,
-      children: [for (final MapEntry(:key, :value) in items.entries) FSelectItem(key, value)],
+      children: [for (final MapEntry(:key, :value) in items.entries) FSelectItem(title: Text(key), value:  value)],
     );
   }
 
@@ -450,7 +450,7 @@ abstract class FSelect<T> extends StatefulWidget with FFormFieldProperties<T> {
               .where((entry) => entry.key.toLowerCase().startsWith(query.toLowerCase()))
               .map((entry) => entry.value)
               .toList(),
-      contentBuilder: (context, _, values) => [for (final value in values) FSelectItem<T>(inverse[value]!, value)],
+      contentBuilder: (context, _, values) => [for (final value in values) FSelectItem<T>(title: Text(inverse[value]!), value: value)],
       searchFieldProperties: searchFieldProperties,
       contentLoadingBuilder: contentLoadingBuilder,
       contentErrorBuilder: contentErrorBuilder,