+++ samples/lib/widgets/select.dart
@@ -33,7 +33,7 @@ class SelectPage extends Sample {
     child: FSelect<String>(
       hint: 'Select a fruit',
       format: (s) => s,
-      children: [for (final fruit in fruits) FSelectItem(fruit, fruit)],
+      children: [for (final fruit in fruits) FSelectItem(title: Text(fruit), value:  fruit)],
     ),
   );
 }
@@ -48,23 +48,23 @@ class DetailedSelectPage extends Sample {
     child: FSelect<String>(
       hint: 'Type',
       format: (s) => s,
-      children: [
-        FSelectItem.from(
-          prefix: const Icon(FIcons.bug),
-          title: const Text('Bug'),
-          subtitle: const Text('An unexpected problem or behavior'),
+      children: const [
+        FSelectItem(
+          prefix: Icon(FIcons.bug),
+          title: Text('Bug'),
+          subtitle: Text('An unexpected problem or behavior'),
           value: 'Bug',
         ),
-        FSelectItem.from(
-          prefix: const Icon(FIcons.filePlus2),
-          title: const Text('Feature'),
-          subtitle: const Text('A new feature or enhancement'),
+        FSelectItem(
+          prefix: Icon(FIcons.filePlus2),
+          title: Text('Feature'),
+          subtitle: Text('A new feature or enhancement'),
           value: 'Feature',
         ),
-        FSelectItem.from(
-          prefix: const Icon(FIcons.messageCircleQuestionMark),
-          title: const Text('Question'),
-          subtitle: const Text('A question or clarification'),
+        FSelectItem(
+          prefix: Icon(FIcons.messageCircleQuestionMark),
+          title: Text('Question'),
+          subtitle: Text('A question or clarification'),
           value: 'Question',
         ),
       ],
@@ -180,8 +180,8 @@ class DividerSelectPage extends Sample {
             for (final item in ['A', 'B']) item: '2$item',
           },
         ),
-        FSelectItem('Level 3', '3'),
-        FSelectItem('Level 4', '4'),
+        const FSelectItem(title: Text('Level 3'), value:  '3'),
+        const FSelectItem(title: Text('Level 4'), value:  '4'),
       ],
     ),
   );
@@ -198,7 +198,7 @@ class SyncSelectPage extends Sample {
       hint: 'Select a fruit',
       format: (s) => s,
       filter: (query) => query.isEmpty ? fruits : fruits.where((f) => f.toLowerCase().startsWith(query.toLowerCase())),
-      contentBuilder: (context, _, fruits) => [for (final fruit in fruits) FSelectItem(fruit, fruit)],
+      contentBuilder: (context, _, fruits) => [for (final fruit in fruits) FSelectItem(title: Text(fruit), value:  fruit)],
     ),
   );
 }
@@ -217,7 +217,7 @@ class AsyncSelectPage extends Sample {
         await Future.delayed(const Duration(seconds: 1));
         return query.isEmpty ? fruits : fruits.where((fruit) => fruit.toLowerCase().startsWith(query.toLowerCase()));
       },
-      contentBuilder: (context, _, fruits) => [for (final fruit in fruits) FSelectItem(fruit, fruit)],
+      contentBuilder: (context, _, fruits) => [for (final fruit in fruits) FSelectItem(title: Text(fruit), value:  fruit)],
     ),
   );
 }
@@ -240,7 +240,7 @@ class AsyncLoadingSelectPage extends Sample {
         padding: const EdgeInsets.all(8.0),
         child: Text('Here be dragons...', style: style.textFieldStyle.contentTextStyle.resolve({})),
       ),
-      contentBuilder: (context, _, fruits) => [for (final fruit in fruits) FSelectItem(fruit, fruit)],
+      contentBuilder: (context, _, fruits) => [for (final fruit in fruits) FSelectItem(title: Text(fruit), value:  fruit)],
     ),
   );
 }
@@ -259,7 +259,7 @@ class AsyncErrorSelectPage extends Sample {
         await Future.delayed(const Duration(seconds: 1));
         throw StateError('Error loading data');
       },
-      contentBuilder: (context, _, fruits) => [for (final fruit in fruits) FSelectItem(fruit, fruit)],
+      contentBuilder: (context, _, fruits) => [for (final fruit in fruits) FSelectItem(title: Text(fruit), value:  fruit)],
       contentErrorBuilder: (context, error, trace) {
         final style = context.theme.selectStyle.iconStyle;
         return Padding(
@@ -289,7 +289,7 @@ class ToggleableSelectPageState extends StatefulSampleState<ToggleableSelectPage
       hint: 'Select a fruit',
       format: (s) => s,
       controller: _controller,
-      children: [for (final fruit in fruits) FSelectItem(fruit, fruit)],
+      children: [for (final fruit in fruits) FSelectItem(title: Text(fruit), value:  fruit)],
     ),
   );
 
@@ -311,7 +311,7 @@ class ClearableSelectPage extends Sample {
       hint: 'Select a fruit',
       format: (s) => s,
       clearable: true,
-      children: [for (final fruit in fruits) FSelectItem(fruit, fruit)],
+      children: [for (final fruit in fruits) FSelectItem(title: Text(fruit), value:  fruit)],
     ),
   );
 }
@@ -333,7 +333,7 @@ class FormatSelectPage extends Sample {
     child: FSelect<({String firstName, String lastName})>(
       hint: 'Select a user',
       format: (user) => '${user.firstName} ${user.lastName}',
-      children: [for (final user in users) FSelectItem(user.firstName, user)],
+      children: [for (final user in users) FSelectItem(title: Text(user.firstName), value:  user)],
     ),
   );
 }
@@ -349,7 +349,7 @@ class ScrollHandlesSelectPage extends Sample {
       hint: 'Select a fruit',
       format: (s) => s,
       contentScrollHandles: true,
-      children: [for (final fruit in fruits) FSelectItem(fruit, fruit)],
+      children: [for (final fruit in fruits) FSelectItem(title: Text(fruit), value:  fruit)],
     ),
   );
 }
@@ -383,7 +383,7 @@ class _FormSelectPageState extends StatefulSampleState<FormSelectPage> with Sing
             hint: 'Select a department',
             format: (s) => s,
             validator: _validateDepartment,
-            children: [for (final department in _departments) FSelectItem(department, department)],
+            children: [for (final department in _departments) FSelectItem(title: Text(department), value:  department)],
           ),
           const SizedBox(height: 25),
           FButton(